requirements.txt
torch>=2.0
iisignature==0.24
numpy
scipy
matplotlib
pyyaml
tqdm
pytest
